Flytta felles storybook ressurser inn i v2/gui/storybook. (#6096)
Slik at vi kan skrive stories for v2/ komponenter som bruker felles ressurser utan å få kompileringsfeil.
