+++ packages/prosess-aarskvantum-oms/src/ÅrskvantumIndex.stories.tsx
@@ -12,8 +12,7 @@ import { Rammevedtak, RammevedtakEnum } from '@k9-sak-web/types/src/omsorgspenge
 import { FraværÅrsakEnum } from '@k9-sak-web/types/src/omsorgspenger/Uttaksperiode';
 import { action } from '@storybook/addon-actions';
 import React from 'react';
-// eslint-disable-next-line import/no-relative-packages
-import alleKodeverk from '../../storybook/stories/mocks/alleKodeverk.json';
+import alleKodeverk from '@k9-sak-web/gui/storybook/mocks/alleKodeverk.json';
 import Aktivitet from './dto/Aktivitet';
 import ÅrskvantumForbrukteDager from './dto/ÅrskvantumForbrukteDager';
 import ÅrskvantumIndex from './ÅrskvantumIndex';
