+++ packages/sak-aktor/src/AktorSakIndex.stories.tsx
@@ -6,8 +6,7 @@ import relasjonsRolleType from '@fpsak-frontend/kodeverk/src/relasjonsRolleType'
 import { fagsakYtelsesType } from '@k9-sak-web/backend/k9sak/kodeverk/FagsakYtelsesType.js';
 import AktorSakIndex from '@k9-sak-web/sak-aktor';
 import { Fagsak } from '@k9-sak-web/types';
-// eslint-disable-next-line import/no-relative-packages
-import alleKodeverk from '../../storybook/stories/mocks/alleKodeverk.json';
+import alleKodeverk from '@k9-sak-web/gui/storybook/mocks/alleKodeverk.json';
 
 const fagsak = {
   saksnummer: '35425245',
