+++ packages/v2/gui/src/storybook/mocks/alleKodeverk.json
