+++ packages/v2/gui/src/storybook/mocks/personopplysninger.ts
@@ -1,0 +1,12 @@
+import type { Personopplysninger } from '../../utils/formidling.js';
+
+/*
+ Separat, minimal versjon av mock data frå gammal kode, kobla til ny type.
+ */
+const personopplysninger: Personopplysninger = {
+  aktoerId: '2821629142423',
+  fnr: '25848395245',
+  navn: 'Asparges Teoretisk',
+};
+
+export default personopplysninger;
